+++ blockpoller/poller.go
@@ -175,7 +175,7 @@ func (p *BlockPoller[C]) run(resolvedStartBlock bstream.BlockRef, stopBlock uint
 			}
 		}
 
-		p.logger.Info("about to fetch block", zap.Uint64("block_to_fetch", blockToFetch), zap.Duration("delay", delay), zap.Bool("keep", false))
+		p.logger.Debug("about to fetch block", zap.Uint64("block_to_fetch", blockToFetch), zap.Duration("delay", delay), zap.Bool("keep", false))
 		if delay != 0 {
 			time.Sleep(delay)
 		}
@@ -213,7 +213,7 @@ func (p *BlockPoller[C]) run(resolvedStartBlock bstream.BlockRef, stopBlock uint
 }
 
 func (p *BlockPoller[C]) processBlock(currentState *cursor, block *pbbstream.Block) (uint64, *string, error) {
-	p.logger.Info("processing block", zap.Stringer("block", block.AsRef()), zap.Uint64("lib_num", block.LibNum), zap.Bool("keep", false))
+	p.logger.Debug("processing block", zap.Stringer("block", block.AsRef()), zap.Uint64("lib_num", block.LibNum), zap.Bool("keep", false))
 	if block.Number < p.forkDB.LIBNum() {
 		panic(fmt.Errorf("unexpected error block %d is below the current LIB num %d. There should be no re-org above the current LIB num", block.Number, p.forkDB.LIBNum()))
 	}
@@ -333,7 +333,7 @@ func (p *BlockPoller[C]) loadNextBlocks(requestedBlock uint64, numberOfBlockToFe
 
 		//only fetch block if it is available on chain
 		if p.blockFetcher.IsBlockAvailable(b) {
-			p.logger.Info("optimistically fetching block", zap.Uint64("block_num", b))
+			p.logger.Debug("optimistically fetching block", zap.Uint64("block_num", b))
 			didTriggerFetch = true
 			nailer.Push(ctx, b)
 		} else {
@@ -362,7 +362,7 @@ func (p *BlockPoller[C]) loadNextBlocks(requestedBlock uint64, numberOfBlockToFe
 }
 
 func (p *BlockPoller[C]) requestBlock(blockNumber uint64, numberOfBlockToFetch int) (*BlockItem, error) {
-	p.logger.Info("requesting block", zap.Uint64("block_num", blockNumber), zap.Bool("keep", false))
+	p.logger.Debug("requesting block", zap.Uint64("block_num", blockNumber), zap.Bool("keep", false))
 
 	lastLog := time.Time{}
 	for {
@@ -395,7 +395,7 @@ func (p *BlockPoller[C]) requestBlock(blockNumber uint64, numberOfBlockToFetch i
 			continue
 		}
 
-		p.logger.Info("block was optimistically polled", zap.Uint64("block_num", blockNumber), zap.Bool("keep", false))
+		p.logger.Debug("block was optimistically polled", zap.Uint64("block_num", blockNumber), zap.Bool("keep", false))
 		return blockItem, nil
 	}
 
@@ -407,7 +407,7 @@ type FetchResponse struct {
 }
 
 func (p *BlockPoller[C]) fetchBlockWithHash(blkNum uint64, hash string) (*pbbstream.Block, error) {
-	p.logger.Info("fetching block with hash", zap.Uint64("block_num", blkNum), zap.String("hash", hash))
+	p.logger.Debug("fetching block with hash", zap.Uint64("block_num", blkNum), zap.String("hash", hash))
 	_ = hash //todo: hash will be used to fetch block from  cache
 
 	p.optimisticallyPolledBlocksLock.Lock()